WIP: XRegs1024 OpenVM integration
- Transpiler extension: decodes 64-bit instructions (all tests pass)
- Circuit u8→u16 for register byte offsets
- Register file expanded to 1024
- RISCV_TO_X86_OVERRIDE_MAP expanded to 1024
- Fibonacci E2E test passes through OpenVM interpreter
- Keccak E2E blocked by call fixup incompatibility (R_RISCV_CALL_PLT
  assumes 4-byte instruction spacing)

+++ extensions/rv32im/circuit/src/auipc/execution.rs
@@ -18,7 +18,7 @@ use crate::common::*;
 #[repr(C)]
 struct AuiPcPreCompute {
     imm: u32,
-    a: u8,
+    a: u16,
 }
 
 impl<A> Rv32AuipcExecutor<A> {
@@ -36,7 +36,7 @@ impl<A> Rv32AuipcExecutor<A> {
         let data: &mut AuiPcPreCompute = data.borrow_mut();
         *data = AuiPcPreCompute {
             imm,
-            a: a.as_canonical_u32() as u8,
+            a: a.as_canonical_u32() as u16,
         };
         Ok(())
     }
